change packagename to lastree for release
