+++ MANIFEST.in
@@ -1,0 +1,2 @@
+prune test
+prune __pycache__
