+++ src/lastree/__init__.py
@@ -1,0 +1,5 @@
+from lastree._version import __version__
+from lastree import utils
+from lastree.utils import set_env, r_script_2_py, tools
+from lastree import main
+
