+++ src/lastree/_version.py
@@ -1,0 +1,2 @@
+import time
+__version__ = str(time.strftime("%d.%m.%Y"))
