--- src/py_interface/__init__.py
@@ -1,4 +1,0 @@
-from . import utils
-from .utils import set_env, r_script_2_py, tools
-
-
