--- src/py_interface/utils/__init__.py
@@ -1,3 +1,0 @@
-from . import tools
-from . import r_script_2_py
-from . import set_env
